back-end_new/requirements.txt
fastapi[all]
uvicorn
pydantic
python-dotenv
google-genai
redis